Move UnsafeMut to simd utils
